# tutorial/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentStatusResponse": {
        "additionalProperties": false,
        "description": "Agent row for dashboards.",
        "properties": {
          "agent_type": {
            "title": "Agent Type",
            "type": "string"
          },
          "avg_task_duration_ms": {
            "title": "Avg Task Duration Ms",
            "type": "number"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_heartbeat_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Heartbeat At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tasks_completed": {
            "title": "Tasks Completed",
            "type": "integer"
          },
          "tasks_failed": {
            "title": "Tasks Failed",
            "type": "integer"
          },
          "uptime_seconds": {
            "title": "Uptime Seconds",
            "type": "number"
          }
        },
        "required": [
          "id",
          "name",
          "agent_type",
          "status",
          "tasks_completed",
          "tasks_failed",
          "avg_task_duration_ms",
          "uptime_seconds",
          "last_heartbeat_at"
        ],
        "title": "AgentStatusResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "The World's First Agentic AI That Learns by Teaching",
    "title": "TUTORIAL API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/agents/": {
      "get": {
        "description": "List registered agents and their persisted status.",
        "operationId": "list_agents_api_v1_agents__get",
        "parameters": [
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentStatusResponse"
                  },
                  "title": "Response List Agents Api V1 Agents  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Agents",
        "tags": [
          "agents"
        ]
      }
    },
...
    "/api/v1/agents/{agent_name}/pause": {
      "post": {
        "description": "Mark an agent offline in persistence (demo pause).",
        "operationId": "pause_agent_api_v1_agents__agent_name__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "title": "Agent Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Pause Agent Api V1 Agents  Agent Name  Pause Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pause Agent",
        "tags": [
          "agents"
        ]
      }
    },
    "/api/v1/agents/{agent_name}/resume": {
      "post": {
        "description": "Mark an agent active again.",
        "operationId": "resume_agent_api_v1_agents__agent_name__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_name",
            "required": true,
            "schema": {
              "title": "Agent Name",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-API-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Resume Agent Api V1 Agents  Agent Name  Resume Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Agent",
        "tags": [
          "agents"
        ]
      }
    },
...
  }
}